tests/wpt/web-platform-tests/xhr/sync-xhr-supported-by-feature-policy.html

<!DOCTYPE html>
<title>Test that sync-xhr is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-feature-policy/#dom-featurepolicy-features">
<link rel="help" href="https://xhr.spec.whatwg.org/#feature-policy-integration">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('sync-xhr', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise sync-xhr.');
</script>
